Fixing failed tests
Test was failing with distance and step count issues.
First step was to set "visited" to true on start node, didn't seem to have
an effect.

Then looked into the test and saw it was hardcoding a start, mid, and
end node, then constructing the path, which only counted 2 steps.
Quick look at the while loop in my function and notice that on breaking the
while loop I forgot to then add the final node that was the start node.

Fixed, built, tests completed successfully.
Runs.